split huge metrics file into multiple files
