+++ scIB/metrics/silhouette.py
@@ -1,0 +1,81 @@
+import pandas as pd
+from sklearn.metrics.cluster import silhouette_score, silhouette_samples
+
+
+def silhouette(
+        adata,
+        group_key,
+        embed,
+        metric='euclidean',
+        scale=True
+):
+    """
+    wrapper for sklearn silhouette function values range from [-1, 1] with 1 being an ideal fit, 0 indicating
+    overlapping clusters and -1 indicating misclassified cells
+
+    :param group_key: key in adata.obs of cell labels
+    :param embed: embedding key in adata.obsm, default: 'X_pca'
+    :param scale: default True, scale between 0 (worst) and 1 (best)
+    """
+    if embed not in adata.obsm.keys():
+        print(adata.obsm.keys())
+        raise KeyError(f'{embed} not in obsm')
+    asw = silhouette_score(
+        X=adata.obsm[embed],
+        labels=adata.obs[group_key],
+        metric=metric
+    )
+    if scale:
+        asw = (asw + 1) / 2
+    return asw
+
+
+def silhouette_batch(
+        adata,
+        batch_key,
+        group_key,
+        embed,
+        metric='euclidean',
+        verbose=True,
+        scale=True
+):
+    """
+    Silhouette score of batch labels subsetted for each group.
+    params:
+        batch_key: batches to be compared against
+        group_key: group labels to be subsetted by e.g. cell type
+        embed: name of column in adata.obsm
+        metric: see sklearn silhouette score
+    returns:
+        all scores: absolute silhouette scores per group label
+        group means: if `mean=True`
+    """
+    if embed not in adata.obsm.keys():
+        print(adata.obsm.keys())
+        raise KeyError(f'{embed} not in obsm')
+
+    sil_all = pd.DataFrame(columns=['group', 'silhouette_score'])
+
+    for group in adata.obs[group_key].unique():
+        adata_group = adata[adata.obs[group_key] == group]
+        n_batches = adata_group.obs[batch_key].nunique()
+        if (n_batches == 1) or (n_batches == adata_group.shape[0]):
+            continue
+        sil_per_group = silhouette_samples(
+            adata_group.obsm[embed],
+            adata_group.obs[batch_key],
+            metric=metric
+        )
+        # take only absolute value
+        sil_per_group = [abs(i) for i in sil_per_group]
+        if scale:
+            # scale s.t. highest number is optimal
+            sil_per_group = [1 - i for i in sil_per_group]
+        d = pd.DataFrame({'group': [group] * len(sil_per_group), 'silhouette_score': sil_per_group})
+        sil_all = sil_all.append(d)
+    sil_all = sil_all.reset_index(drop=True)
+    sil_means = sil_all.groupby('group').mean()
+
+    if verbose:
+        print(f'mean silhouette per cell: {sil_means}')
+    return sil_all, sil_means
